+++ 003_人工智能图像识别领域从业人员的考核标准/readme.md
@@ -47,7 +47,8 @@
     * 3.14 参数初始化细节
     * 3.15 轻量化网络的经验：网络剪枝、网络加速
     * 3.16 搭建ResNet网络的经验
-    * 3.17 ...
+    * 3.17 非线性的表达能力，类似于分段函数
+    * 3.18 ...
 * 雷坤了解8分、理解2分    
 * 雷坤本项总计得分10分
